feat(api): serve MCP through FastAPI
