--- api/__init__.py
@@ -1,1 +1,0 @@
-"""FastAPI application package for agentic-browser."""
