--- api/run.py
@@ -1,14 +1,0 @@
-import uvicorn
-
-
-def run(host: str = "0.0.0.0", port: int = 5454, reload: bool = True):
-    uvicorn.run(
-        "api.main:app",
-        host=host,
-        port=port,
-        reload=reload,
-    )
-
-
-if __name__ == "__main__":
-    run()
